Separate out shadowsocks
